# [Mono.Android] remove IEnumerable iterators called at startup